Migrate UmbracoViewPageTests
